Move auth-bearer to use service TraceProvider.
This PR moves the auth-bearer service to use the service TraceProvider.
